requirements.txt
asgiref==3.10.0
Django==5.2.7
django-js-asset==3.1.2
django-mptt==0.18.0
psycopg2==2.9.11
sqlparse==0.5.3
tzdata==2025.2
